+++ vcfa/sample_vcfa_test_config.json
@@ -67,5 +67,14 @@
       }
     ],
     "rootCertificatePath": "../test-resources/rootCA.pem"
+  },
+  "ldap": {
+    "host": "my.ldap.server.net",
+    "port": 443,
+    "isSsl": true,
+    "username": "admin",
+    "password": "StrongPassword",
+    "baseDistinguishedName": "OU=demo,DC=foo,DC=bar,DC=test,DC=net",
+    "type": "ACTIVE_DIRECTORY"
   }
 }
